Item 7882: Add support for discovery of data involved in transactions through detailed audit logging (#1621)

+++ api/src/org/labkey/api/data/SelectQueryAuditProvider.java
@@ -96,7 +96,7 @@ public class SelectQueryAuditProvider extends AbstractAuditTypeProvider
             Set<PropertyDescriptor> fields = new LinkedHashSet<>();
             fields.add(createPropertyDescriptor("LoggedColumns", PropertyType.STRING));
             fields.add(createPropertyDescriptor("IdentifiedData", PropertyType.STRING, -1));
-            fields.add(createPropertyDescriptor("QueryId", PropertyType.INTEGER));
+            fields.add(createPropertyDescriptor("QueryId", PropertyType.BIGINT));
             _fields = Collections.unmodifiableSet(fields);
         }
